Add rudimentary support for parsing XML with libXML2.
The API is based loosely on that of XOM (http://www.xom.nu/). This is currently very crude but is just sufficient to parse simple XML files and build up a simple DOM-like structure.

Whilst here, remove Whole_Parser.cv (which shouldn't have been imported into the repository in the first place) and make the required changes to the build.