Add heapsort implementation.
Also make 'sort' be a simple front-end to whichever sorting implementation it considers the most efficient given the input data. 'sort' is not guaranteed to be a stable sort.

In the future, a 'stable_sort' function should be added which would be the front-end to the most suitable stable sort.