Add No Default If "ndif" construct.
This is intended to capture the common case where an if statement and its elif clauses are effectively a switch statement whose default action is "shouldn't have got here". The standard case is when there is an enum of sorts and an if statement, each branch of which copes with one of the enums cases. If someone adds another element to the enum, one would prefer that the if statement didn't merrily carry on as if nothing had happened. If "ndif" is used and none of the if braches match an exception is raised.

Therefore the following:

  ndif x:
    ...
  elif y:
    ...

is syntactic sugar for:

  if x:
    ...
  elif y:
    ...
  else:
    <raise exception>

Note that, by definition, it makes no sense for an ndif to have an else clause.

Since this is a common idiom, and its use can help capture bizarre errors early, it seems worth the small increase in syntax.