Be more consistent in recording and using module identifiers.
Module identifiers for .cv files have always supposed to have been the canonicalised path name. However this was something of a bodge job with various bits of code calling File::canon_path on compiler.src_path.

This patch adds in an explicit module_id slot (which is currently the canonicalised path name, but might in the future be changed if that turns out to be useful) to the compiler object which makes using this far more consistent (and resistant to future changes). The src_path slot should probably only be used when presenting errors to the user.