Experimental, rudimentary, floating point number support.
Add a Float class (which is a subtype of Number) and provide basic operations.

Also alter the Int class so that if it performs an operation in which the outcome should be a float then a Float is returned (e.g. add an int and a float, or divide two ints which are not wholly divisible).

Currently there is no direct way to write floats in source code. Use Builtins::Float.new("...") until support is available.