Add simplistic module documentation system.
There is still much to do in terms of quality and quantity of documentation, but this marks a step in the right direction.

The basic idea of the documentation-to-HTML system is simple: .cvd files are an XML format with a few known elements. Any element not understood by the system is assumed to be raw HTML and is passed through unmolested. Clearly this approach is currently rather HTML-centric, but it should be possible to generalise it without too much difficulty.