64 bit support.
This is a major change which goes far deeper than the brief description might suggest. It brings 64 bit support to Converge - both a 64 bit compatible VM, and a compiler which can generate "native" 64 bit .cvb files.

This patch introduces a cross-compilation ability:

  make cross TARGET=x

Will create compiler/convergec.x and compiler/convergel.x for the target x. "make cross-clean TARGET=x" cleans the files associated with a cross-compilation.

The default bootstrapping system is now the 64 bit compiler, since the 64 bit system can easily cross-compile the 32 bit system (but the reverse is not true).

There is a lot to be improved upon in this patch, but it's a big step in the right direction.