Record which variables refer to imports and make them protected.
This patch makes imports slightly special: they define variables that can only be assigned to in one place. What this means is that one can statically know all variables which refer to imports, and one can also know what value has been assigned to them. This is stage #2 in making imports statically calculable.

Violating the variable protection raises a compile-time error. Note that this is not the same as a final variable in Java.

So this will raise an error on the second line:

  import Sys
  import Sys

Just to emphasise that this is quite a deep change, the following will raise a similar error:

  import Sys

  func ():
    nonlocal Sys
	import Sys

as will this:

  import Sys

  func ():
    nonlocal Sys
	Sys := 2