Allow the linker to output Converge executeables which are shell scripts.
This means that one need not necessarily invoke the Converge VM directly. e.g. instead of:

  converge convergel -o x x.cvb
  converge x

one can do:

  converge convergel -s -o x x.cvb
  chmod a+x x
  ./x

In fact, this patch is slightly more generic than it might first appear; the VM is altered so that it allows any arbitrary text to preceed the appearance of the CONVEXEC string in a file. The "real" bytecode is considered to start at the beginning of the first appearance of this string.