Small garbage collector efficiency improvement.
We remove one iteration over all memory chunks by changing the points at which we set the gc_to_be_deleted and gc_incomplete_visit flags. This leads to around a 3% performance increase.

While here, also slightly change the way that memory chunks are shuffled as they are deleted. Previously, when a chunk was deleted, the last chunk in the list overwrote it, thus upsetting the order of the chunks. Now, we shuffle the elements along. This means that at the end of each garbage colllection, all the chunks are still sorted. This doesn't lead to a noticeable performance increase, but it might in the future be possible to exploit this pre-existing sorted data with a more intelligent mergesort-style algorithm.