Beef up error reporting for splices.
This is an analogous patch to the recent beefing-up of quasi-quote error reporting. This patch is not entirely complete, but already catches invalid splices such as:

  $<1>

and:

  func f():
	return [|
      x := 4
	  y := 2
	|]

  ...

  1 + $<f()>

Note that this patch needs to treat splices as special when building ITree's, and to do this neatly, the Converge grammar is changed to formally split the splice and insert rules.