Record the src info of where a normal splice is inserted into.
When performing "normal" splicing, we augment the src infos of the ITree we are splicing in with the position of the splice. This makes "normal" and "DSL" splicing symmetric in the sense that "DSL" splices (while admittedly at the mercy of the DSL implementation function) typically record similar information.

What this means in practise is that for the following example:

  func f():
    return [| 1.foo |]

  $<f()>

the inevitable run-time error will pinpoint both lines 2 and 4 as being related to the error.