Import of multiple modules becomes a conjunction.
Previously this:

  import Strings, Sys

would cause a run-time exception to be raised in the compiler if used as an expression, since a list of imports was returned, whereas the compiler expects expressions to only ever return a single item. Thus when the above is used in an expression it is translated to the following:

  import Strings && import Sys

to avoid this problem. Note that when imports are used as definitions at the top-level of a file, the original behaviour remains.