Ensure that dynamically created new classes have at least one super class.
This isn't an entirely comfortable patch. Since the class "Object" has an empty sequence of super-classes (it's the root class), it's now obvious that it was created by other means during bootstrapping. However, in the interests of preventing people created new classes that have no super-classes (which causes all sorts of confusion), this patch seems reasonable.