Fix how functions and auto-assign to variables works.
This brings some level of sanity to the hodge-podge mechanisms that mean that:

  func f(): ...

is effectively a synonym for:

  f := func f(): ...

unless f is defined as a field in a class, in which case this "auto-assign" doesn't happen.

This patch fixes some conceptual holes, and makes the following work correctly:

  [| func ${"f"} { ... } |]

Note that this patch does not make the same fix for classes; that's for later.