Darwin / Mac OS X x86 port.
This patch contains the minimum number of changes necessary to get Converge working on Darwin x86. This involves slightly rejigging configure.ac, partly to use the helpful pcre-config. The only thing hacked out is the platform specific memmove which Darwin's as refuses to compile, for whatever reason.

Note that there are some errors from use of str[n]cmp as Darwin's gcc; these are harmless, although if someone knows a nice way of shutting it up (that doesn't involve casting static strings to type (u_char *)) then that would be great.

This patch doesn't break Converge on OpenBSD or Darwin, but it should be tested against Cygwin to ensure it still runs.

Thanks to Shin Yoo for giving me access to his box to get this up and running.