Hack around the non-explicitness of 'self' in bound functions.
At various points in the compiler and VM, the fact that 'self' is an implicit variable in bound functions leaks out (e.g. during partial function application). It has a nasty side effect in a corner case of compile-time meta-programming. This patch is a temporary hack to make 'self' explicit at all but the concrete syntax level. This is probably evil. In the long run, I suspect the only good answer is to make users explicitly specify 'self', like Python.