Make 'self' an explicit parameter in bound functions.
self was previously an implicit variable in all bound functions:

  class C:
    func init():
      self.x := 1

In previous versions of Converge, this wasn't a problem, as quasi-quotes did not alpha-rename variables inside functions. Such alpha renaming now does occur (and this is a good thing), but means that it is not possible to dynamically scope 'self' if it is implicit. i.e. this wouldn't work:

  [|
    class C:
      func init():
        &self.x := 1
  |]

This patch means that, like Python, 'self' must be explicitly written as a parameter for bound functions:

  class C:
    func init(self):
      self.x := 1

which means that, like any other variable, self can be dynamically scoped:

  [|
    class C:
      func init(&self):
        &self.x := 1
  |]

This patch does not change the VM, where 'self' has always been encoded as the first parameter of bound functions. Therefore there are other advantages to this patch. For example, partial function application is now more obviously consistent.

This change requires updating the bootstrap compiler.