Avoid arbitrary uses of module lookups.
When module lookups have an "arbitrary" object expression, it becomes impossible to work out what module they refer to. This then defeats the static analysis of module lookups.

This patch first adds a warning message so that such "arbitrary" lookups are reported to the user. As an example this will raise a warning:

  import Sys

  ...

  s := Sys
  s::println(...)

This patch then works around all existing "arbitrary" lookups in the current compiler.