Change the concept of variable definitions and variable lookups to one of "names" and variable lookups.
Previously we distinguished between "variable definitions" and "variable lookups". This was largely to do with the fact that variable would cause ambiguity in the grammar if treated as a single thing defined thus:

  var ::= "ID"
      ::= "&" "ID"
      ::= splice
  	  ::= insert

Then what would the following splice be?

  $<...>

Would it be a top-level splice or a splice of a variable?

However this prevented meaningful reuse of the var grammar rule. We now define the concept of a "name" which, which can be used as a variable definition or elsewhere. This means that variable lookups are now much more explicitly a special case of expressions to ensure the previously mentioned ambiguity is still avoided.