Remove all quasi-quoted code from QQ_Mode.cv.
This removes a world of bootstrapping pain that results every time one tries to alter the fundamentals of QQ_Mode itself. Essentially we strip out some of the automatically generated code that checks boiler-plate stuff, put it into top-level functions in QQ_Mode, and then "lift" function calls to point to this code. This is probably conceptually neater (certainly, it's not any worse), and saves much pain.

As a pleasant side bonus this actually shrinks the size of the compiler by reducing the amount of generated code created by QQ_Mode.