Remove quasi-quoted code from CEI module.
This is a bit icky, but the presence of quasi-quoted code in the CEI module means continual boot-strapping woes, and it's getting quite tedious at a point when there're likely to be several similar changes in the near future.

One day this patch will effectively be backed out, and all will be good.