Change module lookups so that they accept a sequence of lookups.
This makes static analysis of constructs like M:D1:D2 a lot easier. Furthermore we also restrict (at both the concrete and abstract syntax levels) module lookups to be a variable lookup, followed by a sequence of names. Note that module lookups can, if they're quasi-quoted, also be an original name followed by a sequence of names, but that's pretty much the same thing anyway.