Add support for first-class packages.
This means that given a structure like:

  P/
    A.CV

one can now (sometimes) say:

  import P

  P::A

In other words, packages are first-class objects that can be passed around like normal modules. This patch doesn't contain all the needed functionality to be useful, but is an important start. At the moment (and why I said "sometimes"), if P::A wasn't explicitly imported by another part of the program, the P package would not have an A module. Read on to find out why (and be assured that a patch which means that P::A would be included in the above program regardless will be forthcoming).

Packages in Converge are officially unusual. A package only references sub-modules and sub-packages that are used. So if one has:

  P/
    A.cv
	B.cv

and some part of a program imports P::A, but nothing imports P::B, then package P will only appear to have an A component. This is needed to work around a horrible problem that would otherwise happen in compile-time meta-programming in packages. Consider this:

  P/
    A.cv
    B.cv

Imagine that B looks like:

  import B
  $<[| pass |]>

If we compile P/A.cv, and then P/B.cv, what happens? Well, in "old" Converge (which also had first class packages), B.cv needs A.cv, so that's included as the splice is evaluated. But that means constructing the P package, and P contains an A and B components, so the splice would try to incorporate P/B.cvb into the dummy module. Except that file doesn't exist yet, so everything goes splat (even worse, if an *old* P/B.cvb file happens to exist, it will be used, with completely random consequences).

This patch provides an analogous feature to splicing, where only the definitions referenced (directly or indirectly) by a splice are included in a temporary module. Now we only include modules or sub-packages in a package if they're referenced (directly or indirectly) by part of a program.