Make packages include only the minimum necessary set of modules.
This is a big change and is related to the recent patch adding first-class packages and to the more distant patch relating to arbitrary module lookups.

Basically if we have this structure:

  P1/
    A.cv
	B.cv

and a module M.cv:

  import P1

  ...

  P1::f(...)

then the resulting binary will include P1, and P1/A.cv but not P1/B.cv, since it is not referenced. If however A.cv imported B, then B would be included.

This is all predicated on the fact that the names of a modules definitions are fixed at compile-time, and that imports are also "locked" (i.e. one can not assign to import variables).

This patch changes the Converge linker so that it gives an error when a definition which doesn't exist is looked up a module, and it raises a warning when a definition is looked up in an object it is unable to trace. Basically this means that in the above example, we could follow P1::A::B but not P1::A::f::g (because statically we know that P1::A is an import referring to B, but we're not entirely sure what P1::A::f is - it might be a reference to a module, it might not). Entertainingly, this did in fact find one incorrect module lookup in the compiler itself (in an admittedly unused bit of test code).

This patch cures a nasty problem that the old version of Converge suffered from, where compile-time meta-programming in the presence of packages could cause files that were not yet compiled to be imported (resulting in "file not found" errors).