Make builtin modules consistent with user modules.
This is a big patch, touching more of the system than one would ideally like.

The most obvious thing this patch fixes is that builtin modules no longer work by equating definitions with slots. This was evil and meant that:

  import Builtins

  Builtins.Sys

was valid, but trying to access a definition in a user module that way was illegal. A side bonus of this fix is a system speed-up of around 5%.

The second thing this patch addresses is more important, although of less direct interest to most people. Previously it was not possible to discover anything about a builtin module without importing it. Importing is now explicitly seperated out from creating a module. A "created" module (for want of a better term) is a Builtin::Module object which can tell you about its definitions and so on, but which may not have been imported. This makes a lot of things in the bowels of the compiler, in particular, more consistent and safer. To this end, builtin modules need to adopt a consistent naming scheme for their "init" and "import" functions. This patch does that, but does not rename the rest of a modules functions. That's for the next patch.

Because of some of the changes herein, the bootstrap compiler has to be updated.