Make Type_Exception more flexible.
Previously it expected to be given a class as its first argument, but this meant that it could only report errors along the lines of "Expected an X but got a Y". Now one passes an arbitrary string as the first argument, which allows more complex messages to be displayed e.g. "Expected an X or Y but got a Z".