Add initial version of Time module.
The thinking here is that in long term this class will consist of two classes Instant (which denotes a specific point in time) and Range (which denotes a passage of time). These will be semi-opaque classes, which aren't directly instantiated, but are instead created via auxiliary functions which can encode whatever representation Instant and Range use. At the moment we are a little lazy and use the UNIX timespec representation of seconds and nano-seconds, which works well enough.