Add make mode to convergec.
If the -m switch is specified to convergec, then it compiles any files that are out of date (i.e. whose .cvb file does not exist, or which is older than the corresponding .cv file), and then if any updates have been found, automatically creates a CV_Exec file.

This was inspired by a similar feature in GHC.