Rename generator functions to something saner.
Previously it was very hard to work out whether a given function might be a generator or not. Now we try and be a bit more consistent and say that most generators will be prefixed by "iter_". As part of this we rename the "standard" iterate function to "iter", and a couple of other functions too.

Regrettably this requires an update to the bootstrap compiler and linker.