Have the VM determine whether it needs to call 'convergec -m'.
When the VM is called with a source file, and a cached executeable exists, we have the VM quickly work its way through and see if its up to date. This means that execution of an executeable is significantly faster than mucking around creating a pipe, starting up convergec and so on.