Remove the concept of definition quasi-quotes.
Definition quasi-quotes were a hang-over from the Template Haskell influence, that never made it properly into this version of Converge because we now have a more consistent way of saying the same thing. In Converge:

  [d| x := ... |]

is equivalent to:

  [| &x := ... |]

TH doesn't have the second form, hence its need for definition quasi-quotes.

We now simplify things and get rid of all code associated with the former, redundant, concept.