Track module ID imports in module lookups.
This solves a problem where the-concept-previously-know-as-an-original-name could lead to "variable not found" errors. Simplifying things a little (the problem only really crops up in nested quasi-quotes):

  [| CEI::foo |]

would get translated to something like:

  IModule_Lookup(IMod_Id_Import("file.cv"), IVar("foo"))

so CEI might not show up as a free var in quasi-quoted code, and then a "variable not found" error could result if that code was spliced into the file it was created from. Now we explicitly track original names, so that we can add them in as (effectively) free variables when appropriate.