Move the CEI module from the compiler to the library.
Some time ago, all the links from inside the CEI module to the compiler were broken, so it's now possible to move it outside the compiler. This means that we can now use CTMP in the library relatively safely (though using it within the compiler itself is still probably more trouble than it's worth).