Allow the VM to be in bootstrapping mode
Basically the compiler can put the VM into "bootstrapping" mode (this is not the same as the VM's own bootstrapping process), where it translates module IDs between "new" and "old" ids as appropriate.

This makes fun CTMP stuff more practical when compiling the compiler and library. However there is a risk that doing such stuff causes bootstrapping nightmares. Everything under the compiler/ directory should not use CTMP. In the lib/ directory, only files which are used by the compiler should not use CTMP. Therefore the lib Makefile is now explicit about the "minimal" files needed to compiler the compiler. When changing e.g. the bytecode format, only this minimal part of the library need be compiled, and it should be relatively simple to do so.