Refactor XML::Nodes, and add XML::XHTML module.
This makes the naming of the XML nodes more sensible, and gets rid of the unwieldy Text node. Now plain strings can be thrown around instead, which is generally easier to deal with.

The XHTML module is basically just a simple front-end to XML::Nodes. It's notionally intended to be compliant-ish with the XHTML 2.0 draft, but that's not taken too seriously yet.

Because XML::XHTML is the first module to use CTMP with the new bootstrapping system, we have to update the bootstrap compiler.