Obliterate the 'fail' variable.
This has always been an unhappy inheritance from Icon, having never fitted well into the Converge way of doing things. It has gradually been elbowed out, and (outside the VM, where it will unfortunately continue to play a large part), has been reduced to use in the "return fail" idiom. This patch makes "fail" itself a keyword, which causes the current function to return and transmit failure to its caller.

As a side effect of this change, the find/get functions in Dict/Set have to be modified not to take a default value. This is a small change: it was used twice in the codebase, in both cases incorrectly.

This change requires updating the bootstrap compiler.