More build cleanups.
Most importantly, this ends the situation where the placement of "include @abs_top_srcdir@/Makefile.inc" was extremely fragile, and moving it up or down a couple of lines in a Makefile could lead to unpredictable problems. This is achieved by no longer relying on Makefile.inc to define the "%.cvb: %.cv" auto-compilation target.