Beginnings of MinGW support.
For the uninitiated amongst us, MinGW (http://www.mingw.org/) is a sort of "native minimal" version of Cygwin. That means that it has a significantly smaller set of built-in libraries and support, so portability is harder, but the result is not tied in to Cygwin's peculiarities.

This patch does not in any way bring complete support, but it gets Converge close-ish to compiling under MinGW, and that's a useful start. More to come...