Fully separate out module IDs from source paths, and change how packages are dealt with.
This is a monstrous patch without doubt. Converge now fully understands the difference between module IDs and source paths. This finally solves the insanity of case-insensitive filesystems in a sensible way. An important sub-part of this is that packages are now dealt with in a different way. They must be "compiled" so that Converge knows all of the entries of a package, and the correct case of each entry.

This all means that Converge now fully works on MinGW. There is a lot of minor tidying up to be done after this patch, cosmetic variable renaming and so on.

There are two small patches included in this one that would, in an ideal world, be separated out, but which are included as both would require updating the bootstrap compiler and linker. First, the library is moved into lib/Stdlib to make it more consistent with the Compiler library. Second the to_lower_case and to_upper_case functions in strings are renamed to lower_cased and upper_cased respectively.