Add fresh mode compilation to auto-make facility.
What this means is that passing -f to converge or convergec will force any cached files to be ignored, and fresh compilation to be done from scratch. This can be useful when moving files between machines, across Converge versions, or just to test out if an error is a result of stale cached files confusing the compilation process (perhaps caused by change in a machines local clock etc.).