Add the concept of blocks.
Converge has always had two types of "block": function bodies, and quasi quote bodies. This was not very well articulated, and hence these were never well represented in the AST. This is a problem as quasi quotes should also have nonlocal and rename declarations in the same way as functions have. This patch is therefore the first step in solving this issue. It adds the explicit concept of a Block, gives it an AST element, revises the grammar accordingly, and partially unifies functions and quasi-quotes. There are a few rough edges in this patch, but these will be revised appropriately in subsequent patches.