More portable way of determining valid optimisations for GCC versions.
For reasons I won't pretend to understand, on some versions of gcc, piping the content of "-v" through several commands and redirections causes gcc to try to execute ld on a non-existent filename. That causes an error which causes the version not to be reported properly. This version, which is slightly simpler, works around this.