Clean-up of linking code.
Rename link_exec to something more sensible, and also move the serialization part of the code (which isn't really a part of linking) into a more appopriate place.

One thing this patch points out is the slightly random order that arguments are passed to different linking related things. Sometimes it's "(target, ...)", other times "(..., target)". This should be fixed.