The big tabs -> spaces change. Also slightly change the way the default libraries are built.
It's now fairly clear that tabs aren't the way to go (as much as I might like them to be). So this changes all that.

While the bootstrap compiler is being updated, slip in a minor change about where the standard libraries are located. This doesn't have a very visible effect but means that "import Stdlib" finally works.