Add simple, incomplete, test framework for language features plus some initial namespace tests.
This patch starts to bring some semblance of testing to Converge. It defines a (very incomplete) DSL that ultimately should be a good way of testing language features, which often lead to files that (intentionally) won't compile (and thus aren't amenable to more traditional notions of testing). As it is, the handful of tests here were used to test some corner cases with the rename declaration.

This patch also adds a regress target to the main makefile that can be used to automatically run all the valid tests for a particular build of Converge.

Much more needs to be done, but this is a start.