Better printing of error messages.
Previously, particularly in auto-make mode, if you compiled a.cv which caused b.cv to be compiled, then some errors in b.cv (e.g. parsing errors) might appear to have originated in a.cv. Cue confusion. This patch causes convergec to be smarter about the way it prints out error messages so that in such cases it prints file names with error messages if the error is in a different file than the original being compiled. It also tries (in a simplistic way) to make backtrace printouts look the same as convergec error reports.