Convert the repository from darcs to git.
darcs is a wonderful tool in many ways (its user interface is currently unmatched), but its practical issues (including the fact that it often consumes > 600MB of memory doing simple tasks on the Converge repository, effectively DOS'ing the server it's running on) have become too severe.

This patch migrates from darcs to git. Whether git turns out to be the right tool for the job remains to be seen; all that can be said is that darcs definitely isn't the right tool for the job in its current state.