Make .cvl libraries be a complete all-in-one bundle of a package.
A .cvl file is now a bundle of sub-modules and sub-packages all belonging to one package. Not only does this make things a bit clearer, but it means that .cvl files can now be used instead of .cv / .cvb pairs of files. In other words, convergec can look into .cvl files and find imports etc. This means that binary distributions now only need distribute Compiler.cvl and Stdlib.cvl, without bothering about all of the indivdual .cv and .cvb files. This then solves an annoying problem when "converge -f" would over-zealously compile already-compiled files in the standard libraries and so on, and could in some cases cause compilation to fail completely (and in a less than obvious fashion).