Bootstrap the Earley parser.
The parser now parses grammar specifications using itself. This gets rid of most of the horrible old hacky code that previously did this job and (although this patch doesn't do so yet) opens the way for decent reporting of errors in grammars. There is still some ickiness in here (the mk_*brks_map functions spring to mind), but it's a definite improvement over what was here before.