Use ucontext.h when it's available instead of setjmp/longjmp.
In theory this should be a more portable mechanism for those operating systems that support it, as using setjmp/longjmp to jump up the stack frame is not explicitly supported by the standards. However ucontext is probably slower than setjmp/longjmp (on Linux this patch slows things down by 10%). In the future it might be sensible to only use this as a backup option when setjmp/longjmp isn't known to work reliably.