Default to not using ucontext.h unless it's explicitly requested.
It might be portable but it's slow, presumably because it's saving signal handlers and so on. That's an unnecessary cost most of the time for Converge, as signal saving makes sense for threads, but not for continuations (mixing signals and continuations seems like a recipe for disaster, so there's no point explicitly trying to cope with it).