Move backtrace creating functionality over to the new Converge 'backtrace' module.
The previous C version of this was a nightmare to maintain, and very inflexible. This patch means we have a nice Converge module to do the same job, and which we can extend to do nice things in the future.

This also adds the very useful facility that backtraces now include the relevant line of source code (when possible). This makes debugging much more pleasant.

But woe betide anyone who loses Stdlib.cvl though... the VM won't be a happy bunny. There's not much we can do about this.