Record src infos on IInserts.
Previously IInserts had no src infos attached to them (if this was intentional, the intent has since been lost...) which meant that the src info calculation was incorrect, since Converge assumes in several places that every instruction has at least one src info attached to it. This patch means that error messages in files with inserts should now be correct.