Add experimental "if-splice".
If-splices are of the form:

  $if<<cond>>[type]<<expr>>

'cond' is evaluated at compile-time. If it succeeds, 'expr' is evaluated and spliced in. If it fails, an IPass (i.e. the no-op) is substituted. 'type' is as for normal splices (i.e. '' is default, 'c' is capturing, 'p' is pragma).

If-splices finally provide a good way for genuinely conditional compile-time code, since 'expr' need only be syntactically valid if 'cond' fails. In other words, if 'cond' fails, then the variables etc. inside 'expr' are irrelevant. This means, for example, that one can express platform dependent conditionals along the lines of:

  $if<<Platform::Properties == "weirdOS">>c<<[| import WeirdOSMod |]>>
  ...
  $if<<Platform::Properties == "weirdOS">><<[| WeirdOSMod::foo() |]>>

On platforms other than weirdOS, the import of the WeirdOSMod module will never be attempted, nor need that variable even be valid. Previously, this sort of program was impossible to express.

Please note that this an experimental feature and should not be relied upon. It is quite likely to change and evolve.

The idea for this feature was developed during a discussion with Martin Berger.