(Temporary?) hack around the fact that optimisations on GCC 4.2.1 on OpenBSD
The problem is in Strings::join - the str_mem variable can occasionally
revert to a previous value. This appears to be an optimisation (-O or above)
bug, but, weirdly, doing "-O -fno-xxx" for all optimisations doesn't fix it;
and enabling all the optimisations by hand (without using -O) doesn't cause
the bug (but doesn't really seem to optimise the code either). I have no
real idea what is going on...