Prevent registers being clobbered by the stack restoration technique.
In some setups (e.g. gcc on x86_64 with -O2 or above), not all registers are
saved when a function is called. That means that our stack restoration
technique can accidentally clobber registers, giving them the value that
they had on the *previous* call to apply_pump - with consequent hilarity.

What we do here is save all the registers (via setjmp), do all the stack
restoration gubbins, then restore the registers (via longjmp) and return an
object. This prevents (hopefully) clobbering.

This allows us to restore -O3 optimisation for GCC - with a consequent
performance increase of around 1/3.