Pretend that Darwin i386 is really amd64.
When I wasn't looking, OS X made it much more difficult to create i386 only
binaries. With a bit of fiddling, it's possible to create dynamically linked
binaries, but statically linked binaries are extremely painful. Building
everything as 64 bit makes both much easier, so rather than martyr ourselves
in the 32 bit cause, put in a fudge factor to override the fact that this
(apparently) largely 64 bit platform now masquerades as 32 bit in 'uname
-m'.