Import the bare-bones of a PyPy VM.
This patch contains an experimental PyPy VM for Converge. At least
initially, this is meant to be 100% backwards compatible with the C Converge
VM. Currently it is extremely incomplete and can only run toy examples. It
is likely that this VM will change rapidly as a better understanding of the
best way to do this is gained.