Move from a global stack to a per-continuation stack.
This patch moves from a single "giant" stack containing everything to a
global stack of continuation frames and a per-continuation stack of
everything else.

This is to enable a substantial optimisation that PyPy's JIT can make on
fixed size lists. This patch then makes each continuation's stack be a fixed
size. At the moment this is a hack, because we don't know in advance how big
that stack should be: we choose a value that's "probably" OK (i.e. it's
bigger than most reasonable functions seem likely to need). Ideally we need
the compiler to tell us precisely how big this stack should be: that seems
like it should be doable, but it has not yet been done.

Overall, this gives us around a 40% speedup on simple examples relative to
the "giant" stack.