Merge failure frames back into the per-continuation stack.
Previously they were maintained in their own stack; while easy to make
efficient, it confuses visualizing the Converge stacks as failure frames
then appear to be missing.

This patch puts failure frames back into the main stack, and uses a subtly
different caching strategy. There is a teeny slowdown (3-5%), but it's not
significant, and worth the debugging gains. [As a side note, I first tried
putting the cache at the global level, but that lead to a 10% slowdown,
despite my best attempts.]