Add Sys.exit function.
This was a lot more involved than I expected. RPython doesn't appear to give
a neat option for exiting a VM. sys.exit doesn't exist and the SystemExit
exception doesn't carry with it the desired return code. We therefore store
an exit code in the VM, and conspire with main.entry_point to return that to
the enironment. It's all a bit icky, but it does work.