Add slot lookup and apply.
This allows things like o.f() to work. This patch takes a very simple-minded
approach by reusing the existing SLOT_LOOKUP instruction. In the C Converge
VM, this instruction used very expensive path through the VM, so a separate
instruction (the unwieldy CON_INSTR_PRE_SLOT_LOOKUP_APPLY) was added to
shortcut this. It's currently unclear whether this VM will suffer from the
same problem in the long term, so until it's clearly necessary, we stick
with the simple-minded approach, which has the correct semantics, but may
perhaps not be fully optimised.