Basic support for exceptions.
This is a fairly major change and is, I think, the last major remaining "I
wonder how I'll do that in RPython" thing on the list (though there remains
much, much detail to be filled in). This is also a major point of difference
from the C VM which used setjmp/longjmp. This patch mostly piggy-backs off
the Python exception mechanism; fortunately its semantics match Converge's
close enough that it's not too painful.

In terms of exceptions, this isn't the end of the story, but it hopefully
shows that this strategy will work as its expanded upon.