Stop passing the number of parameters to a function on cf.stack.
This has always been a bit of a silly way of doing things; it involves
allocating a Con_Int, then pushing and immediately popping it off of the
stack. Instead, store the number of parameters being passed directly in the
continuation frame. This is a small saving (< 1%), but a saving nonetheless.