Only cache a single failure frame object.
Because failure frames are so frequently allocated, it's faster to maintain a
cache of unused ones than to continually allocate and garbage collect them.
To my initial surprise, the optimum cache size is 1 (rather than
20-something, the size of the cache previously). This makes sense when you
think about the common case of reusing a failure frame from line to line in a
program; reusing a single failure frame captures this common inefficiency
well.

This patch therefore simplifies the VM by storing a single failure frame for
reuse. This leads to an overall speed-up of ~15-20%.