Optimise the size of cf.stack for RPython functions.
Functions internally to the VM don't need much of a cf.stack at all (only for
passing arguments and for storing exception frames). This patch shrinks the
size of such a functions cf.stack to just the size of the number of arguments
passed.

As a hack, while I'm here, reduce the size of Converge cf.stack's to a more
reasonable 32. A function that needs a stack size of 64 is a little way off.

Between the two changes, this leads to a speedup of about 40%.